ai/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "yantra-ai"
version = "0.1.0"
description = "Local-first Yantra AI microservice scaffold."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "groq>=0.31,<1.0",
  "python-dotenv>=1.0,<2.0",
  "sentence-transformers>=3.0,<6.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
voice = [
  "livekit-agents>=1.5,<2.0",
  "livekit-plugins-sarvam>=1.5,<2.0",
  "livekit-plugins-silero>=1.5,<2.0",
]
dev = [
  "pytest>=8.3,<9.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["yantra_ai*"]
